made possible in #994 - now moved
